pyproject.toml
[tool.poetry]
name = "re"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.8"
celery = "^5.3.4"
redis = "^5.0.1"
pytest = "^7.4.2"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
